Clarify uints

+++ contracts/IPullWithdrawable.sol
@@ -1,20 +1,20 @@
 interface IPullWithdrawable {
-    event Withdrawal(address indexed to, uint amount);
-    event WithdrawalERC20(address indexed ERC20Address, address indexed to, uint amount);
+    event Withdrawal(address indexed to, uint256 amount);
+    event WithdrawalERC20(address indexed ERC20Address, address indexed to, uint256 amount);
 
-    function withdraw (uint amount) external;
+    function withdraw (uint256 amount) external;
 
     function withdraw () external;
 
     function withdrawFrom (address payable from) external;
 
-    function withdrawFrom (address payable from, uint amount) external;
+    function withdrawFrom (address payable from, uint256 amount) external;
 
-    function withdrawERC20 (address ERC20Address, uint amount) external;
+    function withdrawERC20 (address ERC20Address, uint256 amount) external;
 
     function withdrawERC20 (address ERC20Address) external;
 
     function withdrawERC20From (address ERC20Address, address payable from) external;
 
-    function withdrawERC20From (address ERC20Address, address payable from, uint amount) external;
+    function withdrawERC20From (address ERC20Address, address payable from, uint256 amount) external;
 }
